Put BBCode in module (#512)
* BBCode in module

* Formatting and BasicBlockCode module name

* Shuffle some names around

* Update src/interpreter/bbcode.jl


* Bump patch version

---------
